requirements.txt
flask==1.1.2
Flask-Script==2.0.6
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
Flask-Login==0.5.0
mysqlclient==2.0.3
redis==2.10.6
unittest2==1.1.0
requests==2.25.1
flasgger==0.9.5
boto3==1.19.2
Jinja2==3.0.2
matplotlib==3.4.3
numpy==1.21.3
pandas==1.3.4
peakUtils==1.3.3